update deps, linting, test cases, etc.

+++ src/fs-helper.ts
@@ -1,14 +1,12 @@
-import * as core from '@actions/core'
-import * as exec from '@actions/exec'
 import * as fs from 'fs'
+import fsExtra from 'fs-extra'
 import * as path from 'path'
 import * as tar from 'tar'
 import * as archiver from 'archiver'
 import * as crypto from 'crypto'
 import * as os from 'os'
-import * as zlib from 'zlib'
 
-export function createTempDir() {
+export function createTempDir(): string {
   const randomDirName = crypto.randomBytes(4).toString('hex')
   const tempDir = path.join(os.tmpdir(), randomDirName)
 
@@ -19,8 +17,10 @@ export function createTempDir() {
   return tempDir
 }
 
-export function removeDir(dir: string) {
-  fs.rmSync(dir, { recursive: true })
+export function removeDir(dir: string): void {
+  if (fs.existsSync(dir)) {
+    fs.rmSync(dir, { recursive: true })
+  }
 }
 
 export interface FileMetadata {
@@ -38,59 +38,90 @@ export async function createArchives(
   const zipPath = path.join(archiveTargetPath, `archive.zip`)
   const tarPath = path.join(archiveTargetPath, `archive.tar.gz`)
 
-  return Promise.all([
-    new Promise<FileMetadata>((resolve, reject) => {
-      const output = fs.createWriteStream(zipPath)
-      const archive = archiver.create('zip')
+  const createZipPromise = new Promise<FileMetadata>((resolve, reject) => {
+    const output = fs.createWriteStream(zipPath)
+    const archive = archiver.create('zip')
 
-      output.on('error', (err: Error) => {
-        reject(err)
-      })
-
-      archive.on('error', (err: Error) => {
-        reject(err)
-      })
-
-      output.on('close', () => {
-        resolve(fileMetadata(zipPath))
-      })
-
-      archive.pipe(output)
-      archive.directory(distPath, false)
-      archive.finalize()
-    }),
-    new Promise<FileMetadata>((resolve, reject) => {
-      const tarStream = tar
-        .c(
-          {
-            file: tarPath,
-            C: distPath, // Change to the source directory for relative paths (TODO)
-            gzip: true
-          },
-          ['.']
-        )
-        .then(() => {
-          resolve(fileMetadata(tarPath))
-        })
-        .catch((err: Error) => reject(err))
+    output.on('error', (err: Error) => {
+      reject(err)
     })
-  ]).then(([zipFile, tarFile]) => ({ zipFile, tarFile }))
+
+    archive.on('error', (err: Error) => {
+      reject(err)
+    })
+
+    output.on('close', () => {
+      resolve(fileMetadata(zipPath))
+    })
+
+    archive.pipe(output)
+    archive.directory(distPath, false)
+    archive.finalize()
+  })
+
+  const createTarPromise = new Promise<FileMetadata>((resolve, reject) => {
+    tar
+      .c(
+        {
+          file: tarPath,
+          C: distPath, // Change to the source directory for relative paths (TODO)
+          gzip: true
+        },
+        ['.']
+      )
+      // eslint-disable-next-line github/no-then
+      .catch(err => {
+        reject(err)
+      })
+      // eslint-disable-next-line github/no-then
+      .then(() => {
+        resolve(fileMetadata(tarPath))
+      })
+  })
+
+  const [zipFile, tarFile] = await Promise.all([
+    createZipPromise,
+    createTarPromise
+  ])
+
+  return { zipFile, tarFile }
 }
 
-export function isDirectory(path: string): boolean {
-  return fs.existsSync(path) && fs.lstatSync(path).isDirectory()
+export function isDirectory(dirPath: string): boolean {
+  return fs.existsSync(dirPath) && fs.lstatSync(dirPath).isDirectory()
 }
 
-export function readFileContents(path: string): Buffer {
-  return fs.readFileSync(path)
+export function readFileContents(filePath: string): Buffer {
+  return fs.readFileSync(filePath)
+}
+
+export function bundleFilesintoDirectory(
+  files: string[],
+  targetDir: string = createTempDir()
+): string {
+  for (const file of files) {
+    if (!fs.existsSync(file)) {
+      throw new Error(`File ${file} does not exist`)
+    }
+
+    if (isDirectory(file)) {
+      const targetFolder = path.join(targetDir, path.basename(file))
+      fsExtra.copySync(file, targetFolder)
+    } else {
+      const targetFile = path.join(targetDir, path.basename(file))
+      fs.copyFileSync(file, targetFile)
+    }
+  }
+
+  return targetDir
 }
 
 // Converts a file path to a filemetadata object by querying the fs for relevant metadata.
-async function fileMetadata(path: string): Promise<FileMetadata> {
-  const stats = fs.statSync(path)
+async function fileMetadata(filePath: string): Promise<FileMetadata> {
+  const stats = fs.statSync(filePath)
   const size = stats.size
   const hash = crypto.createHash('sha256')
-  const fileStream = fs.createReadStream(path)
+  const fileStream = fs.createReadStream(filePath)
   return new Promise((resolve, reject) => {
     fileStream.on('data', data => {
       hash.update(data)
@@ -98,9 +129,9 @@ async function fileMetadata(path: string): Promise<FileMetadata> {
     fileStream.on('end', () => {
       const sha256 = hash.digest('hex')
       resolve({
-        path: path,
-        size: size,
-        sha256: 'sha256:' + sha256
+        path: filePath,
+        size,
+        sha256: `sha256:${sha256}`
       })
     })
     fileStream.on('error', err => {
